// src/app/interfaces/DatoAsociado.ts
export interface DatoAsociado {
    venpas_id: number;
    venpas_idref: string | null;
    venpas_idoriginal: string | null;
    itinerario_id: string;
    ruta_id: string;
    c_origen: string;
    c_destino: string;
    detalle_ruta: string;
    cliente_id: string | null;
    ruc_cliente: string | null;
    razon_social: string | null;
    direccion_cliente: string | null;
    pasajero_id: string;
    documento_pasajero: string;
    nombres_apellidos: string;
    email_pasajero: string;
    telefono_pasajero: string;
    forpag_id: string;
    detalle_forpag: string;
    servicio_id: string;
    detalle_servicio: string;
    tipcom_id: string;
    detalle_tipcom: string;
    tipmov_id: string;
    detalle_tipmov: string;
    tipforpag_id: string;
    detalle_tipforpag: string;
    tarcre_id: string | null;
    detalle_tarcre: string | null;
    c_numboleto: string;
    c_numbolant: string | null;
    n_numasiento: number;
    c_numcontrol: string;
    id_agencia_partida: string;
    direccion_agencia_partida: string;
    fecha_partida: string;
    hora_partida: string;
    id_agencia_llegada: string;
    direccion_agencia_llegada: string;
    fecha_llegada: string;
    hora_llegada: string;
    n_secuencial: number;
    n_imppag: number;
    tipo_tramite: string;
    agencia_id: string;
    nombre_agencia: string;
    usuario_id: string;
    c_login: string;
    canven_id: string;
    detalle_canven: string;
    n_numopeban: string | null;
    d_fecexpres: string | null;
    c_horexpres: string | null;
    n_idaret: number;
    detalle_idaret: string;
    c_rucclicre: string;
    n_esfecabi: number;
    detalle_esfecabi: string;
    c_observaciones: string | null;
    promocion_id: string | null;
    nombre_promocion1: string | null;
    descuento_promocion1: string | null;
    n_ididaret: string;
    audfecins: string;
    audusuins: string;
    audfecmod: string;
    audusumod: string;
    c_estdoc: string;
    c_codref: string | null;
    d_fecvent: string;
    result_ws_id: string | null;
    detalle_result_ws_id: string | null;
    n_esfe: string;
    d_esfe: string;
    n_estado_fe: string | null;
    c_email_contacto: string;
    c_telefono_opcional: string;
    n_info: string;
    correo_encuesta_enviado: number;
    nro_operation_niubiz: string;
    nro_cip_pagoefectivo: string | null;
    nombre_promocion2: string | null;
    descuento_promocion2: string | null;
    correo_enviado1: number;
    correo_enviado2: number;
    venpas_idtx: string | null;
    parentesco_id: string | null;
    venpas_idparents: string | null;
    n_tipo_pasajero: string | null;
    fecha_hora_canje: string;
    fecha_hora_niubiz1: string;
    fecha_hora_niubiz2: string | null;
    fecha_hora_correo: string;
    numboleto_numbolant: string | null;
    anulado_por: string | null;
    estado_order: number;
    pdf_para_descargar: number;
}