// src/app/interfaces/DatosDetallesVenta.ts
export interface DatosDetallesVenta {
    detalle_result_ws_id: string | null;
    numoperacion: string | null;
    canal_venta: string | null;
    boleto: string | null;
    telefono: string | null;
    email: string | null;
    tipo: string | null;
    correo_enviado: number;
    ruc: string | null;
    razon_social: string | null;
    direccion_ruc: string | null;
    numeroCip: string | null;
    fecha_insercion: string | null;
    fecha_hora_niubiz: string | null;
    fecha_hora_niubiz2: string | null;
    fecha_hora_canje: string | null;
    fecha_hora_correo: string | null;
}